@@ -3,7 +3,7 @@
   "description": "Google reCAPTCHA for Nette - Forms",
   "keywords": ["nette", "forms", "captcha", "recaptcha", "google"],
   "type": "library",
-  "license": ["MIT"],
+  "license": "MIT",
   "homepage": "https://github.com/contributte/reCAPTCHA",
   "authors": [
     {
@@ -15,14 +15,14 @@
     "ext-openssl": "To make requests via https"
   },
   "require": {
-    "php": ">= 5.6",
+    "php": ">= 7.1",
     "nette/di": "~2.4.11",
     "nette/forms": "~2.4.7",
-    "nette/utils": "~2.5.1"
+    "nette/utils": "~2.5.3"
   },
   "require-dev": {
-    "ninjify/qa": "^0.4.0",
-    "ninjify/nunjuck": "^0.2.0"
+    "ninjify/qa": "^0.8.0",
+    "ninjify/nunjuck": "^0.2.1"
   },
   "conflict": {
     "nette/http": "<2.4.0"
@@ -37,19 +37,21 @@
       "linter src tests",
       "codesniffer src tests"
     ],
-    "tester": [
+    "tests": [
       "tester -s -p php --colors 1 -C tests/cases"
     ],
     "coverage": [
-      "tester -s -p php --colors 1 -C -d extension=xdebug.so --coverage ./coverage.xml --coverage-src ./src tests/cases"
+      "tester -s -p phpdbg --colors 1 -C --coverage ./coverage.xml --coverage-src ./src tests/cases"
     ],
     "phpstan-install": [
       "mkdir -p temp/phpstan",
-      "composer require -d temp/phpstan phpstan/phpstan:0.8.5",
-      "composer require -d temp/phpstan phpstan/phpstan-nette:0.8.3"
+      "composer require -d temp/phpstan phpstan/phpstan:^0.10",
+      "composer require -d temp/phpstan phpstan/phpstan-deprecation-rules:^0.10",
+      "composer require -d temp/phpstan phpstan/phpstan-nette:^0.10",
+      "composer require -d temp/phpstan phpstan/phpstan-strict-rules:^0.10"
     ],
     "phpstan": [
-      "temp/phpstan/vendor/bin/phpstan analyse -l 1 -c phpstan.neon src"
+      "temp/phpstan/vendor/bin/phpstan analyse -l max -c phpstan.neon src"
     ]
   },
   "extra": {
