@@ -1,30 +1,21 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace Contributte\ReCaptcha\Forms;
 
 use Contributte\ReCaptcha\ReCaptchaProvider;
 use Nette\Forms\Container;
 
-/**
- * @author Milan Felix Sulc <sulcmil@gmail.com>
- */
 final class ReCaptchaBinding
 {
 
-	/**
-	 * @param ReCaptchaProvider $provider
-	 * @param string $name
-	 * @return void
-	 */
-	public static function bind(ReCaptchaProvider $provider, $name = 'addReCaptcha')
+	public static function bind(ReCaptchaProvider $provider, string $name = 'addReCaptcha'): void
 	{
 		// Bind to form container
-		Container::extensionMethod($name, function ($container, $name = 'recaptcha', $label = 'ReCaptcha', $required = TRUE, $message = NULL) use ($provider) {
+		Container::extensionMethod($name, function (Container $container, string $name = 'recaptcha', string $label = 'ReCaptcha', bool $required = true, ?string $message = null) use ($provider): ReCaptchaField {
 			$field = new ReCaptchaField($provider, $label, $message);
 			$field->setRequired($required);
 			$container[$name] = $field;
-
-			return $container[$name];
+			return $field;
 		});
 	}
 
