@@ -1,56 +1,46 @@
-<?php
+<?php declare(strict_types = 1);
 
 namespace Contributte\ReCaptcha;
 
-/**
- * @author Milan Felix Sulc <sulcmil@gmail.com>
- */
 final class ReCaptchaResponse
 {
 
 	// Error code list
-	const ERROR_CODE_MISSING_INPUT_SECRET = 'missing-input-secret';
-	const ERROR_CODE_INVALID_INPUT_SECRET = 'invalid-input-secret';
-	const ERROR_CODE_MISSING_INPUT_RESPONSE = 'missing-input-response';
-	const ERROR_CODE_INVALID_INPUT_RESPONSE = 'invalid-input-response';
-	const ERROR_CODE_UNKNOWN = 'unknow';
+	public const ERROR_CODE_MISSING_INPUT_SECRET = 'missing-input-secret';
+	public const ERROR_CODE_INVALID_INPUT_SECRET = 'invalid-input-secret';
+	public const ERROR_CODE_MISSING_INPUT_RESPONSE = 'missing-input-response';
+	public const ERROR_CODE_INVALID_INPUT_RESPONSE = 'invalid-input-response';
+	public const ERROR_CODE_UNKNOWN = 'unknow';
 
 	/** @var bool */
 	private $success;
 
-	/** @var string */
+	/** @var string[]|string|null */
 	private $error;
 
 	/**
-	 * @param bool $success
-	 * @param string $error
+	 * @param string[]|string|null $error
 	 */
-	public function __construct($success, $error = NULL)
+	public function __construct(bool $success, $error = null)
 	{
-		$this->success = (bool) $success;
+		$this->success = $success;
 		$this->error = $error;
 	}
 
-	/**
-	 * @return boolean
-	 */
-	public function isSuccess()
+	public function isSuccess(): bool
 	{
 		return $this->success;
 	}
 
 	/**
-	 * @return string
+	 * @return string[]|string|null
 	 */
 	public function getError()
 	{
 		return $this->error;
 	}
 
-	/**
-	 * @return string
-	 */
-	public function __toString()
+	public function __toString(): string
 	{
 		return (string) $this->isSuccess();
 	}
